Added ability to add hops to recipe

+++ src/com/biermacht/brews/recipe/Grain.java
@@ -99,7 +99,7 @@ public class Grain extends Ingredient {
 	}
 	
 	@Override 
-	public void setAmount(float amt)
+	public void setAmount(double amt)
 	{
 		setWeight(amt);
 	}
